+++ src/test/java/com/vacaro/benchlab/IntegrationTest.java
@@ -1,8 +1,8 @@
 package com.vacaro.benchlab;
 
 import com.vacaro.benchlab.config.AsyncSyncConfiguration;
-import com.vacaro.benchlab.config.EmbeddedRedis;
 import com.vacaro.benchlab.config.EmbeddedSQL;
+import com.vacaro.benchlab.config.InMemoryCacheTestConfiguration;
 import com.vacaro.benchlab.config.JacksonConfiguration;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -15,8 +15,10 @@ import org.springframework.boot.test.context.SpringBootTest;
  */
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-@SpringBootTest(classes = { BenchLabApp.class, JacksonConfiguration.class, AsyncSyncConfiguration.class })
-@EmbeddedRedis
+@SpringBootTest(
+    properties = "spring.cache.type=simple",
+    classes = { BenchLabApp.class, JacksonConfiguration.class, AsyncSyncConfiguration.class, InMemoryCacheTestConfiguration.class }
+)
 @EmbeddedSQL
 public @interface IntegrationTest {
 }