+++ src/test/java/com/vacaro/benchlab/config/InMemoryCacheTestConfiguration.java
@@ -1,0 +1,17 @@
+package com.vacaro.benchlab.config;
+
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.cache.CacheManager;
+import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.cache.concurrent.ConcurrentMapCacheManager;
+import org.springframework.context.annotation.Bean;
+
+@TestConfiguration(proxyBeanMethods = false)
+@EnableCaching
+public class InMemoryCacheTestConfiguration {
+
+    @Bean
+    CacheManager cacheManager() {
+        return new ConcurrentMapCacheManager();
+    }
+}